Moved some functions from gameplay to gameplay_utils and modified game configuration to showcase full capabilities @WiktorB2004 #3
