+++ include/gameplay.h
@@ -57,37 +57,6 @@ typedef struct
 } Game;
 
 /*!
- * \brief This function initializes regular game of size (3x3).
- *
- * This function initializes game of size 3x3 by allocating needed memory
- * and other struct values.
- *
- * @param board Pointer to board to initialize.
- */
-void initialize_Board(Board *board);
-
-/*!
- * \brief This function initializes game of specified size (NxN).
- *
- * This function initializes game of specified size by allocating needed memory
- * and other struct values.
- *
- *
- * @param game Pointer to Game to initialize.
- * @param size The size of the game to initialize
- */
-void initialize_Game(Game *game, int size);
-
-/*!
- * \brief This function deallocates memory of the Game.
- *
- * This function deallocates the memory allocated in initialize_Game function.
- *
- * @param game Pointer to the game to free.
- */
-void free_Game(Game *game);
-
-/*!
  * \brief This function places the move in specified row/column.
  *
  * This function places the move using provided row and column on 3x3 board.
