+++ src/gameplay/gameplay.c
@@ -6,59 +6,6 @@
 
 // FIXME: Create memory deallocation functions - use them inside engine after the end of the game
 
-void initialize_Board(Board *board)
-{
-    board->status = IN_PROGRESS;
-    board->moves_count = 0;
-    for (int i = 0; i < 3; i++)
-    {
-        for (int j = 0; j < 3; j++)
-        {
-            board->value[i][j] = EMPTY;
-        }
-    }
-}
-
-void initialize_Game(Game *game, int size)
-{
-    game->board = (Board **)malloc(sizeof(Board *) * (size * size));
-    if (game->board == NULL)
-    {
-        fprintf(stderr, "Memory allocation for Game failed.\n");
-        exit(EXIT_FAILURE);
-    }
-    // FIXME: Make it random
-    for (int i = 0; i < (size * size); i++)
-    {
-        game->board[i] = (Board *)malloc(sizeof(Board));
-        initialize_Board(game->board[i]);
-    }
-    game->turn = X;
-    game->board_size = size;
-    game->moves_count = 0;
-    game->status = NOT_STARTED;
-}
-
-void free_Game(Game *game)
-{
-    if (game == NULL)
-    {
-        return;
-    }
-
-    if (game->board != NULL)
-    {
-        for (int i = 0; i < (game->board_size * game->board_size); i++)
-        {
-            if (game->board[i] != NULL)
-            {
-                free(game->board[i]);
-            }
-        }
-        free(game->board);
-    }
-}
-
 void modify_board(Board *board, int row, int column, Player player)
 {
     switch (board->status)
@@ -151,18 +98,22 @@ void gameplay(Game *game)
     {
         // -------------TO BE DELETED--------------------
         // Sketch of printing the board
-        for (int i = 0; i < 3; i++)
+        for (int k = 0; k < game->board_size; k++)
         {
-            for (int j = 0; j < 3; j++)
+            for (int i = 0; i < 3; i++)
             {
-                printf("%d", game->board[0]->value[i][j]);
+                for (int j = 0; j < 3; j++)
+                {
+                    printf("%d", game->board[k]->value[i][j]);
+                }
+                printf("\n");
             }
             printf("\n");
         }
         // Sketch of getting user input
         if (game->board_size > 1)
         {
-            printf("Select the board to place a move (0 - %d)\n", game->board_size);
+            printf("Select the board to place a move (0 - %d)\n", game->board_size - 1);
             scanf("%d", &mv_board);
         }
         printf("Enter the point to place a move (0 - %d) (0 - %d)\n", 2, 2);
