+++ utils/gameplay_utils.c
@@ -1,6 +1,61 @@
+#include <stdlib.h>
+#include <stdio.h>
 #include <stdbool.h>
 #include "../include/gameplay.h"
 #include "../include/utils/gameplay_utils.h"
+
+void initialize_Board(Board *board)
+{
+    board->status = IN_PROGRESS;
+    board->moves_count = 0;
+    for (int i = 0; i < 3; i++)
+    {
+        for (int j = 0; j < 3; j++)
+        {
+            board->value[i][j] = EMPTY;
+        }
+    }
+}
+
+void initialize_Game(Game *game, int size)
+{
+    game->board = (Board **)malloc(sizeof(Board *) * (size * size));
+    if (game->board == NULL)
+    {
+        fprintf(stderr, "Memory allocation for Game failed.\n");
+        exit(EXIT_FAILURE);
+    }
+    // FIXME: Make it random
+    for (int i = 0; i < (size * size); i++)
+    {
+        game->board[i] = (Board *)malloc(sizeof(Board));
+        initialize_Board(game->board[i]);
+    }
+    game->turn = X;
+    game->board_size = size;
+    game->moves_count = 0;
+    game->status = NOT_STARTED;
+}
+
+void free_Game(Game *game)
+{
+    if (game == NULL)
+    {
+        return;
+    }
+
+    if (game->board != NULL)
+    {
+        for (int i = 0; i < (game->board_size * game->board_size); i++)
+        {
+            if (game->board[i] != NULL)
+            {
+                free(game->board[i]);
+            }
+        }
+        free(game->board);
+    }
+}
 
 bool check_board_horizontally(Board *board)
 {
